Fix integer signedness warnings in C++ code
- Change loop variables from int to size_t in fims_math.hpp (2 locations)
- Change loop variables from int to size_t in rcpp_models.hpp (3 locations)
- Change loop variables from int to size_t in rcpp_population.hpp (1 location)
- Change loop variables from int to size_t in rcpp_distribution.hpp (3 locations)
- Add static_cast<size_t> for n_years.get() comparison in rcpp_recruitment.hpp

+++ inst/include/common/fims_math.hpp
@@ -335,7 +335,7 @@ inline const Type ad_max(const Type &a, const Type &b, Type C = 1e-5) {
 template <class T>
 T sum(const std::vector<T> &v) {
   T ret = 0.0;
-  for (int i = 0; i < v.size(); i++) {
+  for (size_t i = 0; i < v.size(); i++) {
     ret += v[i];
   }
   return ret;
@@ -352,7 +352,7 @@ T sum(const std::vector<T> &v) {
 template <class T>
 T sum(const fims::Vector<T> &v) {
   T ret = 0.0;
-  for (int i = 0; i < v.size(); i++) {
+  for (size_t i = 0; i < v.size(); i++) {
     ret += v[i];
   }
   return ret;
