+++ inst/include/interface/rcpp/rcpp_objects/rcpp_models.hpp
@@ -829,19 +829,19 @@ class CatchAtAgeInterface : public FisheryModelInterfaceBase {
     double of_value = Rcpp::as<double>(report["objective_function_value"]);
 
     fims::Vector<double> gradient(grad.size());
-    for (int i = 0; i < grad.size(); i++) {
+    for (size_t i = 0; i < grad.size(); i++) {
       gradient[i] = grad[i];
     }
     // Assume grouped_out is an Rcpp::List
     std::map<std::string, std::vector<double>> grouped_cpp;
     if (grouped_out.size() > 0) {
       Rcpp::CharacterVector names = grouped_out.names();
-      for (int i = 0; i < grouped_out.size(); i++) {
+      for (size_t i = 0; i < grouped_out.size(); i++) {
         std::string key = Rcpp::as<std::string>(names[i]);
         Rcpp::NumericVector vec =
             grouped_out[i];  // each element is a numeric vector
         std::vector<double> vec_std(vec.size());
-        for (int j = 0; j < vec.size(); j++) {
+        for (size_t j = 0; j < vec.size(); j++) {
           vec_std[j] = vec[j];
         }
         grouped_cpp[key] = vec_std;
